@@ -1,54 +1,27 @@
 {
-  "name": "chew",
-  "version": "0.1.0",
+  "name": "template.inte",
   "private": true,
+  "version": "0.0.0",
+  "type": "module",
   "scripts": {
-    "serve": "vue-cli-service serve",
-    "build": "vue-cli-service build",
-    "lint": "vue-cli-service lint"
+    "dev": "vite",
+    "build": "vite build",
+    "preview": "vite preview"
   },
   "dependencies": {
-    "autoprefixer": "^10",
-    "core-js": "^3.8.3",
-    "dayjs": "^1.11.4",
-    "postcss": "^8",
-    "prismjs": "^1.28.0",
-    "tailwindcss": "^3",
-    "vue": "^2.6.14",
-    "vue-prism-editor": "^1.3.0",
-    "vue-router": "^3.5.1"
+    "@heroicons/vue": "^2.0.18",
+    "dayjs": "^1.11.13",
+    "ky": "^1.0.1",
+    "prismjs": "^1.29.0",
+    "vue": "^3.3.4",
+    "vue-prism-editor": "^2.0.0-alpha.2",
+    "vue-router": "^4.2.4"
   },
   "devDependencies": {
-    "@babel/core": "^7.12.16",
-    "@babel/eslint-parser": "^7.12.16",
-    "@vue/cli-plugin-babel": "~5.0.0",
-    "@vue/cli-plugin-eslint": "~5.0.0",
-    "@vue/cli-plugin-router": "~5.0.0",
-    "@vue/cli-service": "~5.0.0",
-    "axios": "^0.26.1",
-    "eslint": "^7.32.0",
-    "eslint-plugin-vue": "^8.0.3",
-    "vue-cli-plugin-axios-x": "~0.1.5",
-    "vue-cli-plugin-tailwind": "~3.0.0",
-    "vue-template-compiler": "^2.6.14"
-  },
-  "eslintConfig": {
-    "root": true,
-    "env": {
-      "node": true
-    },
-    "extends": [
-      "plugin:vue/essential",
-      "eslint:recommended"
-    ],
-    "parserOptions": {
-      "parser": "@babel/eslint-parser"
-    },
-    "rules": {}
-  },
-  "browserslist": [
-    "> 1%",
-    "last 2 versions",
-    "not dead"
-  ]
+    "@vitejs/plugin-vue": "^4.2.3",
+    "autoprefixer": "^10.4.15",
+    "postcss": "^8.4.29",
+    "tailwindcss": "^3.3.3",
+    "vite": "^4.4.5"
+  }
 }
