@@ -1,11 +1,3 @@
 <template>
-  <div>
-    <router-view />
-  </div>
+  <router-view />
 </template>
-
-<script>
-export default {
-    name: 'LayoutIndex',
-}
-</script>
