@@ -1,62 +1,46 @@
-import Vue from 'vue';
-import VueRouter from 'vue-router';
-
-Vue.use(VueRouter);
+import { createRouter, createWebHashHistory } from "vue-router";
 
 const routes = [
   {
-    path: '/',
-    component: () => import('@/layouts/DefaultLayout/Index.vue'),
+    path: "/",
+    component: () => import("../layouts/DefaultLayout/Index.vue"),
     children: [
       {
-        path: '/',
-        name: 'home',
-        component: () => import('@/views/HomeView.vue'),
+        path: "",
+        component: () => import("../views/HomeView.vue"),
       },
       {
-        path: 'profile',
-        name: 'profile',
-        component: () => import('@/views/ProfileView.vue'),
+        path: "writer",
+        component: () => import("../views/WriterView.vue"),
       },
       {
-        path: 'writer',
-        name: 'writer',
-        component: () => import('@/views/WriterView.vue'),
-      },
-      {
-        path: 'code',
-        name: 'code',
-        component: () => import('@/views/CodeView.vue'),
+        path: '/result/:gumId',
+        component: () => import('../views/ResultView.vue'),
         props: true,
       },
     ]
   },
   {
-    path: '/coder',
-    name: 'code-writer',
-    component: () => import('@/views/CoderView.vue'),
+    path: "/coder",
+    component: () => import("../views/CoderView.vue"),
   },
   {
-    path: '/r',
-    component: () => import('@/layouts/DefaultLayout/Index.vue'),
+    path: "/r",
+    component: () => import("../layouts/DefaultLayout/Index.vue"),
     children: [
       {
-        path: ':code',
-        name: 'reader',
-        component: () => import('@/views/ReaderView.vue'),
+        path: ":gumId",
+        component: () => import("../views/ReaderView.vue"),
         props: true,
       },
     ]
-  },
-  {
-    path: '*',
-    name: 'not-found',
-    component: () => import('@/views/NotFoundView.vue'),
   },
 ];
 
-const router = new VueRouter({
-  routes
+const router = createRouter({
+  scrollBehavior: () => ({ top: 0 }),
+  history: createWebHashHistory(),
+  routes,
 });
 
 export default router;
